# Optimization: send premultiplied alpha from vertex to fragment shader.