--- a/internal/graphicsdriver/metal/graphics_darwin.go
+++ b/internal/graphicsdriver/metal/graphics_darwin.go
@@ -232,8 +232,11 @@ struct FragmentShaderImpl {
       c.rgb /= c.a + (1.0 - sign(c.a));
       c = (color_matrix_body * c) + color_matrix_translation;
       c.rgb *= c.a;
+      c *= v.color;
+      c.rgb = min(c.rgb, c.a);
+    } else {
+      c *= v.color;
     }
-    c *= v.color;
     return c;
   }
 };