@divVerent 2022-02-23
Check failures were fun:

>   Could not connect to azure.archive.ubuntu.com:80 (52.250.76.244), connection timed out [IP: 52.250.76.244 80]