@divVerent 2022-02-23
Benchmark results of 01-start-to-hub.dem on Dell E6500 (the underpowered GPU with the large screen, shader exec time bound):

- before 352fd76312fab3f11e369cb8ef26af333814c433: 171.7 seconds
- before 352fd76312fab3f11e369cb8ef26af333814c433 + this change: 171.6 seconds (yes, basically no change)
- main at 352fd76312fab3f11e369cb8ef26af333814c433: 119.3 seconds
- main at 352fd76312fab3f11e369cb8ef26af333814c433 + this change:  92.9 seconds (22% faster!)

Why does this change only really take effect with the nearest shader? Because it's really about fillrate - my game runs at 640x360 and has about 2x overdraw at the settings used here, so it renders 461k pixels - however this gets scaled up to 1920x1080 -> 2M pixels. This final render pass really seems to be what matters.