@divVerent 2022-02-23
I just found the condition under which a < 0 can break things, and added a comment as well as adapted the check.

The part I missed was that these colors are unpremultiplied, so my arguments failed epically - a sign flips when premultiplying which can wreak havoc.

The proper condition is r >= 0 && g >= 0 && b >= 0 && r <= 1 && g <= 1 && b <= 1 && a >= 0, and the only possible condition that allows even more requires analyzing the vertex colors.

OTOH, if this condition holds, then multiplying this with the vertex colors does not make the vertex colors any more pathological than they already might be, essentially meaning that the clamping can never kick in due to the vertex colors themselves being unpremultiplied too.

Negative matrix colors can combine with negative vertex colors to something positive, which can cause clamping. It might be possible that I actually just need r >= -1, etc. - did not analyze close enough, probably not worthwile.