@hajimehoshi 2022-02-23
I'm confused. Why don't we need the condition like `r <= a`?