@divVerent 2022-02-23
Because colors in this layer (atlas.Image) are all unpremultiplied, and they previously got premultiplied in the fragment shader (now will in the vertex shader - this is what I primarily am moving here, as well as making the min() optional).