@hajimehoshi 2022-02-23
> So a >= 0 and r <= 1 here implies r <= a in the fragment shader.

What about the scale (r, g, b, a) = (0.5, 0.5, 0.5, 0) for the color (1, 1, 1, 1)? Will the result be (0.5, 0.5, 0.5, 0)?

(For `min`, I want to be flexible about the compatibility, but I'd like to know what is kept and what is not exactly.)