@divVerent 2022-02-23
If the scale is (r, g, b, a) = (0.5, 0.5, 0.5, 0), this

- now: becomes (0, 0, 0, 0) in the vertex shader, and invariably makes the fully-transparent in the fragment shader
- previously: stayed (0.5, 0.5, 0.5, 0) in the vertex shader, got turned into (0, 0, 0, 0) by the fragment shader, and invariably makes the output fully-transparent

This is what is to be expected about the unpremultiplied color (0.5, 0.5, 0.5, 0), which is "fully transparent grey". The fact that it is grey is entirely invisible, and should be irrelevant, when alpha is zero - and premultiplying like Ebiten does ensures precisely that.