@divVerent 2022-02-23
Note that there is one _real_ difference caused by changing the code to do color interpolation in domain of premultiplied rather than unpremultiplied colors - one can construct examples of vertex colorings that look rather different in the interpolated area.

However, in almost all cases the previous behavior would actually be considered a bug, as it makes the "color" of alpha-zero vertices visible.

As an example, change the "shapes" example as follows:

```
diff --git a/examples/shapes/main.go b/examples/shapes/main.go
index 57b0da4c..2854d07b 100644
--- a/examples/shapes/main.go
+++ b/examples/shapes/main.go
@@ -112,16 +112,21 @@ func rect(x, y, w, h float32, clr color.RGBA) ([]ebiten.Vertex, []uint16) {
        x1 := x + w
        y1 := y + h
 
+       r = 0
+       g = 0.25
+       b = 0
+       a = 1
+
        return []ebiten.Vertex{
                {
                        DstX:   x0,
                        DstY:   y0,
                        SrcX:   1,
                        SrcY:   1,
-                       ColorR: r,
-                       ColorG: g,
-                       ColorB: b,
-                       ColorA: a,
+                       ColorR: 1,
+                       ColorG: 0,
+                       ColorB: 0,
+                       ColorA: 0,
                },
                {
                        DstX:   x1,
@@ -157,8 +162,6 @@ func rect(x, y, w, h float32, clr color.RGBA) ([]ebiten.Vertex, []uint16) {
 }
 
 func (g *Game) Update() error {
-       g.count++
-       g.count %= 240
        return nil
 }
 
```

Note how I made one vertex of the quads "fully transparent bright red", and all other vertices "fully opaque dark green".

Before: http://www.rm.cloudns.org/img/uploaded/e4fcea908c3a22d0b32864343fbe42a0.png

After: http://www.rm.cloudns.org/img/uploaded/8c5d93a08ad1a36d1ad76d0f32668542.png

A human, asked to interpolate between these (a transparent piece of glass and a solid dark green piece of paper) would sure agree with then new rendering. That red tint that only occurs between vertices, but never at vertices, is the wellknown artifact that comes from interpolating unpremultiplied colors, and I assume it is why Ebiten usually thinks in premultiplied colors internally.