@divVerent 2022-02-23
Should also mention that the _common_ cases of vertex interpolation are unchanged by this:

- Same color, different alpha - both ways of doing the math result in the same color tone at the same alpha slope.
- Different color, sample alpha - both ways of doing the math result in the same alpha level with the same color slope.