@hajimehoshi 2022-02-23
> now: becomes (0, 0, 0, 0) in the vertex shader, and invariably makes the fully-transparent in the fragment shader

OK acknowledged.