@hajimehoshi 2022-02-23
> A human, asked to interpolate between these (a transparent piece of glass and a solid dark green piece of paper) would sure agree with then new rendering. That red tint that only occurs between vertices, but never at vertices, is the wellknown artifact that comes from interpolating unpremultiplied colors, and I assume it is why Ebiten usually thinks in premultiplied colors internally.

I agree. The new version is much better and makes sense. My understanding is that this is a very similar issue of wrong interpolation due to straight-alpha (non-premultiplied-alpha) colors.

(To be exact, interpolation might have to consider gamma correction, but this is a different issue)