@divVerent 2022-02-23
Yes, it is precisely the same interpolation issue you also get with images in straight-alpha notation when scaling them component-wise without regard for the alpha channel. This can be fixed in a scaling algorithm, of course, however the ones commonly used on GPUs do not.

The only difference is that we had the same issue here in vertex domain rather than texel domain. But indeed - in general, colors should be premultiplied before interpolation.

This underlying issue also affects Kage shaders which I am not fixing here as the chance is higher to break underlying code - this requires further analysis. I do wonder if we are lucky enough to find that nobody uses Fragment()'s "color vec4" argument in Kage with alpha != 1 anyway and thus we can just change it to be premultiplied. Maybe worth tracking this related item as an issue?